requirements.txt
flask
line-bot-sdk==3.13.0
orjson
gunicorn